btlp: the boards, and the symbol that was hiding a dozen matches
Six units come out of btlp_66, all of them the same shape - a table of four
records, a position, and the two calls that grow them into place and shrink
them away again. func_800A7DE0 and func_800A7F68 are BtlBoardOpen and
BtlBoardShut now, and include/persona/btlp/board.h carries the eight bytes a
record is described by, the same eight a party marker takes.

  boards.c        seven open/shut pairs. Four have no caller left in the
                  image; they are in it, so they are here.
  menuboards.c    the orders, tactics, config and status pages, and the
                  tactics board's five member lines - name or EMPTY, and the
                  column standing for the order that member is under.
  choiceprompt.c  the two-option prompt, both sets of it.
  pickgrid.c      the eighteen cells a target is picked on, built from one
                  template with its script table rewritten between calls, and
                  the five cursors that sit on the party's own squares.
  markershow.c    putting a member's marker back up, and the two sweeps that
                  decide who wants one.
  memberboards.c  the five member boards along the bottom. The close matches;
                  the open is the right instructions in the wrong registers
                  and is left to the permuter.

Three levers did most of the work and are worth keeping in mind. Reading an
array by index rather than walking a pointer of your own is what puts the
derived pointer after the hoisted constants in the preheader, which is where
the image has it - it is what took every one of the choice prompt's four
routines from 93% to exact. A value that varies with the outer loop is
written as the expression, not carried in a variable, for the same reason.
And an if whose arms are laid out the wrong way round is written the other
way: the image branches to the arm it puts last.

g_btl_offer_answers was never a table of answers. It sits at 0x8005D6AC,
which is Char.list inside actors[0], and while it stood there splat could not
resolve any reference past +0x58 into a party record - every one came out as
a fresh D_8005D6B8 instead of the g_btl_actors+0x64 the C emits, and counted
as a difference. The table BtlOfferAnswer actually reads is
g_btl_demon_talk_profiles. Dropping the name moved eighteen routines to exact
without their sources being touched.

objflat comes out of its guard: BtlDrawObjFlat was already byte-exact and
only its expected object was stale.

ovl.btlp 16.74% -> 18.29%, 297 matched functions -> 335, and no category
down. All eleven executables still build byte-exact.

+++ src/btlp/moodretire.c
@@ -57,7 +57,7 @@ void BtlMoodRetire(void)
         i++;
     } while (i < BTL_MOODS);
 
-    BtlPanelSetImage(g_btl_panel_gauges != 0, g_btl_panel_gauges);
+    BtlPanelSetImage((short)g_btl_panel_gauges != 0, g_btl_panel_gauges);
     BtlSetMoodGauges(g_btl_offer[g_btl_offer_slot].mood[0],
                      g_btl_offer[g_btl_offer_slot].mood[1],
                      g_btl_offer[g_btl_offer_slot].mood[2],
@@ -66,4 +66,3 @@ void BtlMoodRetire(void)
 #else
 INCLUDE_ASM("btlp/nonmatchings/moodretire", BtlMoodRetire);
 #endif
-
